pyproject.toml
[project]
name = "hospital-construction-builders"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "pillow>=10.0.0",
    "streamlit>=1.28.0",
]
